Modify styling for the Login Page, Register Page, User Home Page
Modified the styling for the Login Page, Register Page, User Home Page
as well as the View Recipes page. The changes were needed to make the
pages look consistent with each other.
There is still work to be done styling the buttons and trimming the
margins so that the horizontal rule fits them.
It may be a possibility to have the horizontal rule to go across the page
instead.