Modify styling for the User Home Page
Modified styling for the User Home Page. The buttons were still
set to the default styling of Bootstrap and the browser.
I wanted custom styling for the buttons as I felt using Bootstrap's
plug and play feature would not allow me to come back and revise
the styling in case the page changes.