Add the Create Recipe page and styling
Added the Create Recipe page and its styling.
It differs from the mockup because I was unable to find a way to
implement the design I had envisioned. I have instead opted for
modifying a form to suit my needs.

Also included in this commit are modifications to the input
elements of the Login and Register pages.
Added logout links to all the pages that need them.
Removed a reference to a JavaScript file from Bootstrap.
Sourcing JavaScript from Bootstrap's CDN instead.
Modified styling for the User Home page.